Time: 217 ms (5.21%), Space: 81.1 MB (14.22%) - LeetHub

+++ 2537-count-the-number-of-good-subarrays/2537-count-the-number-of-good-subarrays.cpp
@@ -3,7 +3,7 @@ public:
     long long countGood(vector<int>& nums, int k) {
         map<int,int>mp;
         int n=nums.size();
-        int res=0,i=0,j=0;
+        long long  res=0,i=0,j=0;
         int pair=0;
         while(j<n){
             pair+=mp[nums[j]];
